backend/requirements.txt

fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
faiss-cpu==1.7.4
numpy==1.24.3
pypdf2==3.0.1
python-docx==1.1.0
python-pptx==0.6.23
aiofiles==23.2.1
httpx==0.25.1
tiktoken==0.5.1

